Convert craco config to typescript
