+++ package.json
@@ -8,7 +8,7 @@
   "dependencies": {
     "@audius/libs": "1.2.54",
     "@audius/stems": "0.3.20",
-    "@craco/craco": "6.4.2",
+    "@craco/craco": "6.4.3",
     "@hcaptcha/react-hcaptcha": "0.3.6",
     "@optimizely/optimizely-sdk": "4.0.0",
     "@project-serum/sol-wallet-adapter": "0.2.5",
@@ -186,6 +186,7 @@
     "@types/amplitude-js": "8.0.2",
     "@types/body-scroll-lock": "2.6.1",
     "@types/classnames": "2.2.10",
+    "@types/craco__craco": "6.4.0",
     "@types/hls.js": "0.13.0",
     "@types/jest": "24.9.1",
     "@types/js-yaml": "3.12.4",
@@ -202,6 +203,7 @@
     "@types/react-native-push-notification": "3.0.8",
     "@types/react-redux": "7.1.7",
     "@types/react-router-dom": "5.3.3",
+    "@types/webpack": "4.41.32",
     "@typescript-eslint/eslint-plugin": "^2.30.0",
     "@typescript-eslint/parser": "2.30.0",
     "abort-controller": "3.0.0",
